02_monitoring/web-app/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.10
Django==4.0.1
djangorestframework==3.13.1
djongo==1.3.6
elasticsearch==7.16.2
idna==3.3
psycopg2-binary==2.9.3
pymongo==3.12.1
pytz==2021.3
requests==2.27.1
sqlparse==0.2.4
urllib3==1.26.7
